clean up a bunch of server errors and start of crud macro

+++ packages/rust/wasm/conundrum_wasm/src/features/queries/emojis.rs
@@ -12,8 +12,8 @@ pub fn search_conundrum_emojis_in_docs_container(query: String,
                                                  per_page: usize)
                                                  -> Result<JsValue, JsValue> {
     let res = search_emojis_to_docs_container(query,
-                                              Some(PaginationParams { per_page: per_page as u32,
-                                                                      page: page as u32 }))
+                                              Some(PaginationParams { per_page: per_page as u64,
+                                                                      page: page as u64 }))
         .map_err(|x| {
                                                   serde_wasm_bindgen::to_value(&x).unwrap_or_default()
         })?;
